requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
scikit-learn==1.8.0
pandas>=2.2.2
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.122.0
requests==2.31.0
